refactor: 모듈 이름 변경, 디렉토리 추가 #2
