+++ src/chat/chat.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
-import { ChatRoomController } from './chat-room.controller';
-import { ChatRoomService } from './chat-room.service';
+import { ChatRoomController } from './chat.controller';
+import { ChatRoomService } from './chat.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ChatRoom } from './entities/chat-room.entity';
-import { ChatRoomRepository } from './chat-room.repository';
+import { ChatRoomRepository } from './repositories/chat-room.repository';
 
 @Module({
 	imports: [TypeOrmModule.forFeature([ChatRoom])],
