+++ src/chat/enums/chat-room-type.enum.ts
